+++ fluxplaneplugin/datasource.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 	"strings"
 
-	coredatasource "github.com/fluxplane/fluxplane-core/core/datasource"
 	"github.com/fluxplane/fluxplane-core/orchestration/pluginhost"
+	coredatasource "github.com/fluxplane/fluxplane-datasource"
 
 	dex "github.com/fluxplane/fluxplane-dex"
 )